siem: replace O(N) LatencyTracker sample buffer with circular index (Issue #1044)
LatencyTracker.Record previously shifted the entire sample buffer left on every
write when the buffer was full: copy(lt.samples, lt.samples[1:]). At 10k events/
sec with a 1000-sample buffer this is a 1000-element memmove on every Record
call, making the hot write path O(N) rather than O(1).

Replaced the growing []float64 slice with a fixed [1000]float64 array and a
circular index: Record writes to lt.samples[lt.head%1000] and increments head.
count = min(count+1, 1000) tracks how many slots hold valid data. No heap
allocation or copy on any Record call.

Extracted percentileFromSorted(sorted []float64, p float64) as a private helper
so GetStats can sort the snapshot buffer once and call percentileFromSorted three
times (p50/p95/p99) rather than invoking GetPercentile three times with three
independent sort passes under three separate read-lock acquisitions.

GetPercentile and GetStats both copy lt.samples[0:lt.count] into a local slice
before releasing the read lock, then sort and compute outside the lock. This
keeps write latency low regardless of sort cost.

NewLatencyTracker no longer performs the make([]float64, 0, 1000) heap
allocation; the struct is now zero-value ready.

Also fixed two pre-existing data races in siem_engine.go exposed by the updated
test suite: GetMetrics was doing a non-atomic struct copy (metricsCopy := *se.metrics)
while concurrent workers used atomic.Add* on the same fields without holding
metricsLock. Fixed by building the snapshot field-by-field and using atomic.Load*
for all atomically-updated counters. updateMetrics had the same pattern reading
TotalEntriesProcessed plainly under the write lock; changed to atomic.LoadInt64.
Fixed a pre-existing uint64 underflow in TestSIEMEngine_MemoryUsage where GC
freeing more memory than the baseline caused wraparound; changed to int64 arithmetic.

Tests added:
- TestLatencyTracker_CircularBuffer_OldestDropped: records 1500 samples, verifies
  p0 reflects 501ms (oldest retained) not 1ms (first ever written)
- TestLatencyTracker_Percentiles_KnownDistribution: records 1-100ms monotonic
  distribution, verifies p50=50.5/p95=95.05/p99=99.01 and that GetStats returns
  identical values without re-sorting via GetPercentile
- TestLatencyTracker_ConcurrentRecord: 20 goroutines * 200 writes under -race,
  verifies count cap and no data races

Fixes #1044

+++ features/siem/siem_engine_test.go
@@ -756,11 +756,13 @@ func TestSIEMEngine_MemoryUsage(t *testing.T) {
 	var memStatsAfter runtime.MemStats
 	runtime.ReadMemStats(&memStatsAfter)
 
-	memoryGrowth := memStatsAfter.Alloc - memStatsBefore.Alloc
+	// Use int64 arithmetic to handle the case where GC frees more than was
+	// allocated since the baseline (uint64 underflow would give a false huge value).
+	memoryGrowth := int64(memStatsAfter.Alloc) - int64(memStatsBefore.Alloc)
 	t.Logf("Memory growth: %d bytes (%.2f MB)", memoryGrowth, float64(memoryGrowth)/(1024*1024))
 
 	// Verify memory growth is reasonable (less than 50MB for this test)
-	assert.Less(t, memoryGrowth, uint64(50*1024*1024),
+	assert.Less(t, memoryGrowth, int64(50*1024*1024),
 		"Excessive memory growth: %d bytes", memoryGrowth)
 }
